Added a generic types version of DomainException. Added a generic types version of DomainEvent. It is no longer possible to create a Reference or a VersionedReference with an abstract type.
Fixed a bug that caused events to remain in the changes collection even after a rollback call.
Fixed a bug that resulted in a generic reference being attached to exceptions raised by the LoadFromHistory method of EventCentricAggregateRoot.
Fixed a bug that resulted in a generic reference being attached to DomainEvents created using the aggregate instance variation of the constructor.